Use unsigned long long for node counting.
As Craig will tell you, it's really annoying when your node count goes
negative after two weeks...

+++ back2school/c++/mwcz.h
@@ -17,7 +17,7 @@ using weight_type = unsigned long;
  */
 class MWCZ {
     public:
-        long nodes;
+        unsigned long long nodes;
         weight_type max_weight;
         milliseconds cpu_time;
         int n;
